+++ src/Plugin.Fingerprint.Abstractions/FingerprintImplementationBase.cs
@@ -5,25 +5,28 @@ namespace Plugin.Fingerprint.Abstractions
 {
     public abstract class FingerprintImplementationBase : IFingerprint
     {
-        public Task<FingerprintAuthenticationResult> AuthenticateAsync(string reason, CancellationToken cancellationToken = default(CancellationToken))
+        public Task<FingerprintAuthenticationResult> AuthenticateAsync(string reason, CancellationToken cancellationToken = default)
         {
             return AuthenticateAsync(new AuthenticationRequestConfiguration(reason), cancellationToken);
         }
 
-        public async Task<FingerprintAuthenticationResult> AuthenticateAsync(AuthenticationRequestConfiguration authRequestConfig, CancellationToken cancellationToken = new CancellationToken())
+        public async Task<FingerprintAuthenticationResult> AuthenticateAsync(AuthenticationRequestConfiguration authRequestConfig, CancellationToken cancellationToken = default)
         {
-            if(!await IsAvailableAsync())
+            if(!await IsAvailableAsync(authRequestConfig.AllowAlternativeAuthentication))
                 return new FingerprintAuthenticationResult { Status = FingerprintAuthenticationResultStatus.NotAvailable };
 
             return await NativeAuthenticateAsync(authRequestConfig, cancellationToken);
         }
 
-        public async Task<bool> IsAvailableAsync()
+        public async Task<bool> IsAvailableAsync(bool allowAlternativeAuthentication = false)
         {
-            return await GetAvailabilityAsync() == FingerprintAvailability.Available;
+            return await GetAvailabilityAsync(allowAlternativeAuthentication) == FingerprintAvailability.Available;
         }
 
-        public abstract Task<FingerprintAvailability> GetAvailabilityAsync();
+        public abstract Task<FingerprintAvailability> GetAvailabilityAsync(bool allowAlternativeAuthentication = false);
+
+        public abstract Task<AuthenticationType> GetAuthenticationTypeAsync();
+
         protected abstract Task<FingerprintAuthenticationResult> NativeAuthenticateAsync(AuthenticationRequestConfiguration authRequestConfig, CancellationToken cancellationToken);
     }
 }