+++ src/Plugin.Fingerprint.Android/Contract/IAuthenticationFailedListener.cs
@@ -3,5 +3,6 @@ namespace Plugin.Fingerprint.Contract
     public interface IAuthenticationFailedListener
     {
         void OnFailedTry();
+        void OnHelp(FingerprintAuthenticationHelp help, string nativeHelpText);
     }
 }