+++ src/Plugin.Fingerprint.NotImplemented/FingerprintImplementation.cs
@@ -6,9 +6,14 @@ namespace Plugin.Fingerprint
 {
     internal class FingerprintImplementation : FingerprintImplementationBase
     {
-        public override async Task<FingerprintAvailability> GetAvailabilityAsync()
+        public override Task<FingerprintAvailability> GetAvailabilityAsync(bool allowAlternativeAuthentication = false)
         {
-            return FingerprintAvailability.NoImplementation;
+            return Task.FromResult(FingerprintAvailability.NoImplementation);
+        }
+
+        public override Task<AuthenticationType> GetAuthenticationTypeAsync()
+        {
+            return Task.FromResult(AuthenticationType.None);
         }
 
         protected override Task<FingerprintAuthenticationResult> NativeAuthenticateAsync(AuthenticationRequestConfiguration authRequestConfig, CancellationToken cancellationToken = new CancellationToken())