+++ src/Plugin.Fingerprint.iOS/FingerprintImplementation.cs
@@ -28,7 +28,8 @@ namespace Plugin.Fingerprint
             Tuple<bool, NSError> resTuple;
             using (cancellationToken.Register(CancelAuthentication))
             {
-                resTuple = await _context.EvaluatePolicyAsync(LAPolicy.DeviceOwnerAuthenticationWithBiometrics, authRequestConfig.Reason);
+                var policy = GetPolicy(authRequestConfig.AllowAlternativeAuthentication);
+                resTuple = await _context.EvaluatePolicyAsync(policy, authRequestConfig.Reason);
             }
 
             if (resTuple.Item1)
@@ -37,65 +38,76 @@ namespace Plugin.Fingerprint
             }
             else
             {
-                switch ((LAStatus)(int)resTuple.Item2.Code)
+                // #79 simulators return null for any reason
+                if (resTuple.Item2 == null)
                 {
-                    case LAStatus.AuthenticationFailed:
-                        var description = resTuple.Item2.Description;
-                        if (description != null && description.Contains("retry limit exceeded"))
-                        {
-                            result.Status = FingerprintAuthenticationResultStatus.TooManyAttempts;
-                        }
-                        else
-                        {
-                            result.Status = FingerprintAuthenticationResultStatus.Failed;
-                        }
-                        break;
-
-                    case LAStatus.UserCancel:
-                    case LAStatus.AppCancel:
-                        result.Status = FingerprintAuthenticationResultStatus.Canceled;
-                        break;
-
-                    case LAStatus.UserFallback:
-                        result.Status = FingerprintAuthenticationResultStatus.FallbackRequested;
-                        break;
-
-                    case LAStatus.TouchIDLockout:
-                        result.Status = FingerprintAuthenticationResultStatus.TooManyAttempts;
-                        break;
-
-                    default:
-                        result.Status = FingerprintAuthenticationResultStatus.UnknownError;
-                        break;
+                    result.Status = FingerprintAuthenticationResultStatus.UnknownError;
+                    result.ErrorMessage = "";
                 }
-
-                result.ErrorMessage = resTuple.Item2.LocalizedDescription;
+                else
+                {
+                    result = GetResultFromError(resTuple.Item2);
+                }
             }
 
             CreateNewContext();
             return result;
         }
 
-        public override async Task<FingerprintAvailability> GetAvailabilityAsync()
+        public override async Task<FingerprintAvailability> GetAvailabilityAsync(bool allowAlternativeAuthentication = false)
         {
-            NSError error;
-
             if (_context == null)
                 return FingerprintAvailability.NoApi;
 
-            if (_context.CanEvaluatePolicy(LAPolicy.DeviceOwnerAuthenticationWithBiometrics, out error))
+            var policy = GetPolicy(allowAlternativeAuthentication);
+            if (_context.CanEvaluatePolicy(policy, out var error))
                 return FingerprintAvailability.Available;
 
             switch ((LAStatus)(int)error.Code)
             {
-                case LAStatus.TouchIDNotAvailable:
+                case LAStatus.BiometryNotAvailable:
                     return FingerprintAvailability.NoSensor;
-                case LAStatus.TouchIDNotEnrolled:
+                case LAStatus.BiometryNotEnrolled:
                 case LAStatus.PasscodeNotSet:
                     return FingerprintAvailability.NoFingerprint;
                 default:
                     return FingerprintAvailability.Unknown;
             }
+        }
+
+        public override async Task<AuthenticationType> GetAuthenticationTypeAsync()
+        {
+            if (_context == null)
+                return AuthenticationType.None;
+
+            // we need to call this, because it will always return none, if you don't call CanEvaluatePolicy
+            var availibility = await GetAvailabilityAsync(false);
+
+            // iOS 11+
+            if (_context.RespondsToSelector(new Selector("biometryType")))
+            {
+                switch (_context.BiometryType)
+                {
+                    case LABiometryType.None:
+                        return AuthenticationType.None;
+                    case LABiometryType.TouchId:
+                        return AuthenticationType.Fingerprint;
+                    case LABiometryType.FaceId:
+                        return AuthenticationType.Face;
+                    default:
+                        return AuthenticationType.None;
+                }
+            }
+
+            // iOS < 11
+            if (availibility == FingerprintAvailability.NoApi ||
+                availibility == FingerprintAvailability.NoSensor || 
+                availibility == FingerprintAvailability.Unknown)
+            {
+                return AuthenticationType.None;
+            }
+
+            return AuthenticationType.Fingerprint;
         }
 
         private void SetupContextProperties(AuthenticationRequestConfiguration authRequestConfig)
@@ -104,12 +116,59 @@ namespace Plugin.Fingerprint
             {
                 _context.LocalizedFallbackTitle = authRequestConfig.FallbackTitle;
             }
-#if !__MAC__ // will be included in Cycle 9
+
             if (_context.RespondsToSelector(new Selector("localizedCancelTitle")))
             {
                 _context.LocalizedCancelTitle = authRequestConfig.CancelTitle;
             }
-#endif
+        }
+
+        private LAPolicy GetPolicy(bool allowAlternativeAuthentication)
+        {
+            return allowAlternativeAuthentication ?
+                LAPolicy.DeviceOwnerAuthentication :
+                LAPolicy.DeviceOwnerAuthenticationWithBiometrics;
+        }
+
+        private FingerprintAuthenticationResult GetResultFromError(NSError error)
+        {
+            var result = new FingerprintAuthenticationResult();
+
+            switch ((LAStatus)(int)error.Code)
+            {
+                case LAStatus.AuthenticationFailed:
+                    var description = error.Description;
+                    if (description != null && description.Contains("retry limit exceeded"))
+                    {
+                        result.Status = FingerprintAuthenticationResultStatus.TooManyAttempts;
+                    }
+                    else
+                    {
+                        result.Status = FingerprintAuthenticationResultStatus.Failed;
+                    }
+                    break;
+
+                case LAStatus.UserCancel:
+                case LAStatus.AppCancel:
+                    result.Status = FingerprintAuthenticationResultStatus.Canceled;
+                    break;
+
+                case LAStatus.UserFallback:
+                    result.Status = FingerprintAuthenticationResultStatus.FallbackRequested;
+                    break;
+
+                case LAStatus.TouchIDLockout:
+                    result.Status = FingerprintAuthenticationResultStatus.TooManyAttempts;
+                    break;
+
+                default:
+                    result.Status = FingerprintAuthenticationResultStatus.UnknownError;
+                    break;
+            }
+
+            result.ErrorMessage = error.LocalizedDescription;
+
+            return result;
         }
 
         private void CancelAuthentication()
@@ -139,8 +198,8 @@ namespace Plugin.Fingerprint
             if (!info.IsOperatingSystemAtLeastVersion(minVersion))
                 return;
 #else
-			if (!UIDevice.CurrentDevice.CheckSystemVersion(8, 0))
-				return;
+            if (!UIDevice.CurrentDevice.CheckSystemVersion(8, 0))
+                return;
 #endif
             // Check LAContext is not available on iOS7 and below, so check LAContext after checking iOS version.
             if (Class.GetHandle(typeof(LAContext)) == IntPtr.Zero)