+++ src/Sample/SMS.Fingerprint.Sample/MainView.xaml.cs
@@ -9,31 +9,48 @@ namespace SMS.Fingerprint.Sample
     public partial class MainView : ContentPage
     {
         private CancellationTokenSource _cancel;
+        private bool _initialized;
 
         public MainView()
         {
             InitializeComponent();
         }
 
+        protected override async void OnAppearing()
+        {
+            base.OnAppearing();
+
+            if (!_initialized)
+            {
+                _initialized = true;
+                lblAuthenticationType.Text = "Auth Type: " + await Plugin.Fingerprint.CrossFingerprint.Current.GetAuthenticationTypeAsync();
+            }
+        }
+
         private async void OnAuthenticate(object sender, EventArgs e)
         {
-            _cancel = swAutoCancel.IsToggled ? new CancellationTokenSource(TimeSpan.FromSeconds(10)) : new CancellationTokenSource();
-            lblStatus.Text = "";
-            var result = await Plugin.Fingerprint.CrossFingerprint.Current.AuthenticateAsync("Prove you have fingers!", _cancel.Token);
-
-            await SetResultAsync(result);
+            await AuthenticateAsync("Prove you have fingers!");
         }
 
         private async void OnAuthenticateLocalized(object sender, EventArgs e)
         {
+            await AuthenticateAsync("Beweise, dass du Finger hast!", "Abbrechen", "Anders!", "Viel zu schnell!");
+        }
+
+        private async Task AuthenticateAsync(string reason, string cancel = null, string fallback = null, string tooFast = null)
+        {
             _cancel = swAutoCancel.IsToggled ? new CancellationTokenSource(TimeSpan.FromSeconds(10)) : new CancellationTokenSource();
             lblStatus.Text = "";
 
-            var dialogConfig = new AuthenticationRequestConfiguration("Beweise, dass du Finger hast!")
-            {
-                CancelTitle = "Abbrechen",
-                FallbackTitle = "Anders!"
+            var dialogConfig = new AuthenticationRequestConfiguration(reason)
+            { // all optional
+                CancelTitle = cancel,
+                FallbackTitle = fallback,
+                AllowAlternativeAuthentication = swAllowAlternative.IsToggled
             };
+
+            // optional
+            dialogConfig.HelpTexts.MovedTooFast = tooFast;
 
             var result = await Plugin.Fingerprint.CrossFingerprint.Current.AuthenticateAsync(dialogConfig, _cancel.Token);