+++ src/sanity-check/MainActivity.cs
@@ -1,0 +1,30 @@
+﻿using System;
+using Android.App;
+using Android.App.Job;
+using Android.Widget;
+using Android.OS;
+using Android.Views;
+
+namespace AndroidResourceSanity
+{
+    [Activity(Label = "AndroidResourceSanity", MainLauncher = true, Icon = "@mipmap/icon")]
+    public class MainActivity : Activity
+    {
+        int count = 1;
+
+        protected override void OnCreate(Bundle savedInstanceState)
+        {
+            base.OnCreate(savedInstanceState);
+
+            // Set our view from the "main" layout resource
+            SetContentView(Resource.Layout.Main);
+
+            // Get our button from the layout resource,
+            // and attach an event to it
+            Button button = FindViewById<Button>(Resource.Id.myButton);
+
+            button.Click += delegate { button.Text = string.Format("{0} clicks!", count++); };
+        }
+    }
+}
+