+++ src/sanity-check/Properties/AssemblyInfo.cs
@@ -1,18 +1,20 @@
-﻿using System.Resources;
-using System.Reflection;
+﻿using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+using Android.App;
 
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("SMS.Fingerprint.Sample")]
+[assembly: AssemblyTitle("AndroidResourceSanity")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("SMS.Fingerprint.Sample")]
-[assembly: AssemblyCopyright("Copyright 2015")]
+[assembly: AssemblyProduct("AndroidResourceSanity")]
+[assembly: AssemblyCopyright("Copyright ©  2017")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
-[assembly: NeutralResourcesLanguage("en")]
+[assembly: ComVisible(false)]
 
 // Version information for an assembly consists of the following four values:
 //
@@ -23,6 +25,6 @@ using System.Reflection;
 //
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.4.3")]
-[assembly: AssemblyVersion("1.4.3")]
-[assembly: AssemblyFileVersion("1.4.3")]
+// [assembly: AssemblyVersion("1.4.6")]
+[assembly: AssemblyVersion("1.4.6")]
+[assembly: AssemblyFileVersion("1.4.6")]